Moved login/sign up into components folder and added background to buttons
