+++ frontend/src/Components/LoginForm/LogInForm.js
@@ -1,0 +1,21 @@
+import React from 'react';
+import { Link } from "react-router-dom"
+import "./../../Pages/SignUpLogIn/SignUpLogIn.css";
+
+function LogIn(props) {
+    return (
+        <div>
+            <div className="nav-buttons">
+                    <Link id="loginLink" to="/login"><button style={{backgroundColor: "#979797"}} className="lsButton" onClick={props.clearError}>Login</button></Link>
+                    <Link id="signUpBtn" to="/signup"><button className="lsButton" onClick={props.clearError}>Sign Up</button></Link>
+                </div>
+            <form id="signUpForm">
+                <input className="form-control form-control-lg signUpInput" type="text" name="email" placeholder="Email" onChange={props.handleChange} /><br />
+                <input className="form-control form-control-lg signUpInput" type="password" name="password" onChange={props.handleChange} placeholder="Password" /><br />
+                <input className="btn signUpSubmit" type="button" value="Log In" onClick={props.submitForm} />
+            </form>
+        </div>
+    )
+}
+
+export default LogIn;
